make Mixer::play take Sound, not Tape
Get rid of two play() calls visible to the user, and the intermediate
type and that binds a playable chunk and its associated filter
parameters, so that a mixer's play now takes the sound itself and the
parameters to be passed to the sound's play function.

Probably too hifi though; the type arguments are getting nasty. Just
testing, haha.