add collision groups, fix gear collisions
Specify categories and collision bits for added bodies so that the world
does not interfere with the gears, so we don't need to cut holes in the
world where the moving gears are.

The gears' positions do not need to be readjusted each time (I think) so
remove the code for that.